backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI/ML
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
pinecone-client==2.2.4
chromadb==0.4.18

# Data processing
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2

# ML libraries
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
datasets==2.14.7

# Integrations
simple-salesforce==1.12.6
authlib==1.2.1
onelogin==2.0.0

# Utilities
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
boto3==1.29.7

# Monitoring and logging
structlog==23.2.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

